feat: ✨ track fx on table or chart dataSource
