+++ apps/builder/src/widgetLibrary/Chart/widgetConfig.tsx
@@ -1,8 +1,6 @@
 import { v4 } from "uuid"
 import { ReactComponent as ChartWidgetIcon } from "@/assets/widgetCover/chart.svg"
 import i18n from "@/i18n/config"
-import { getActionList } from "@/redux/currentApp/action/actionSelector"
-import store from "@/store"
 import { WidgetConfig } from "@/widgetLibrary/interface"
 
 export const CHART_WIDGET_CONFIG: WidgetConfig = {
